requirements.txt
# Core PyTorch dependencies - compatible with CUDA 12.4
torch>=2.0.0
torchvision>=0.15.0

# Computer vision and image processing
opencv-python
scikit-image
imageio-ffmpeg

# Scientific computing
numpy
scipy
matplotlib

# Video processing
av

# Utility libraries
addict
einops
future
pyyaml
requests
timm
yapf

# OCR functionality
rapidocr-onnxruntime
onnxruntime-gpu